+++ site-cookbooks/alterf/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of the alterf cookbook.
 
+## 1.6.0
+- Add package install for development to default recipe
+- Open ports for tcp
+
 ## 1.5.0
 - Remove recipes created in denebola