+++ site-cookbooks/regulus/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of the regulus cookbook.
 
+## 1.5.0
+- Add package install for development to default recipe
+- Open ports for tcp
+
 ## 1.4.0
 - Remove files created in zosma